check bevore creating config file

+++ install.sh
@@ -1,11 +1,15 @@
 #!/usr/bin/env sh
 
-sudo mkdir -p /etc/server-updater
-cat << EOF | sudo tee /etc/server-updater/server-updater.env > /dev/null
-WEBHOOK_URL_INFO=
-WEBHOOK_URL_ERROR=
-EOF
-sudo chmod 006 /etc/server-updater/server-updater.env
+if [ ! -f /tmp/foo.txt ]; then
+    sudo mkdir -p /etc/server-updater
+    cat << EOF | sudo tee /etc/server-updater/server-updater.env > /dev/null
+    WEBHOOK_URL_INFO=
+    WEBHOOK_URL_ERROR=
+    EOF
+    sudo chmod 006 /etc/server-updater/server-updater.env
+
+    echo "created config file at /etc/server-updater/server-updater.env"
+fi
 
 LATEST_VERSION="$(curl -sL https://github.com/0xFEEDC0DE-dev/server-updater/releases \
                 | grep -o 'releases/tag/[0-9]*.[0-9]*.[0-9]*' \
@@ -19,5 +23,4 @@ sudo chmod 007 /usr/local/bin/server-updater
 echo "0 2 * * * mkdir -p /tmp/server-updater && /usr/local/bin/server-updater --env-file /etc/server-updater/server-updater.env >> /tmp/server-updater/out.log 2>> /tmp/server-updater/err.log" | sudo crontab -u root -
 
 echo "Downloaded binary version ${LATEST_VERSION} to /usr/local/bin/server-updater"
-echo "Put config file at /etc/server-updater/server-updater.env"
 echo "Created Cronjob to start every night"
